+++ src/Exception/UnknownLoginProviderException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Drupal\os2forms_form_login\Exception;
+
+/**
+ * Unknown login provider exception.
+ */
+class UnknownLoginProviderException extends \RuntimeException {
+
+}